+++ src/max_div/_core/_math/__init__.py
@@ -1,2 +1,2 @@
 from .powers_of_2 import _POWERS_OF_2_F32_MAX_EXP, _POWERS_OF_2_F32_MIN_EXP, power_of_2_f32
-from .select_k_minmax import select_k_max, select_k_min
+from .select_k_minmax import select_k_max, select_k_max_masked, select_k_min